Automate / reassign multi-appliance tests
